Update climate.py
If AC is found open through the remote control then assign AUTO to it
If the hvac_mode is not set correctly, it may cause Homekit to encounter a situation where the device becomes unresponsive to control commands. To prevent this, if we detect that the device has been turned on remotely, we will assign a default state to it.